vision_model_options.py
#!/usr/bin/env python3
"""
Vision Model Options for MATLAB RAG

This file documents different vision model options you could use
to improve image understanding beyond the current BLIP model.
"""

VISION_MODEL_OPTIONS = {
    "current": {
        "name": "BLIP-2 Opt-2.7B (UPGRADED)",
        "model_id": "Salesforce/blip2-opt-2.7b",
        "strengths": [
            "Highly detailed technical descriptions",
            "Visual Question Answering (VQA) capabilities",
            "Multi-strategy analysis with technical prompts",
            "Excellent understanding of plots and diagrams",
            "MATLAB-specific visualization recognition"
        ],
        "limitations": [
            "Larger model size (2.7B parameters)",
            "Higher memory requirements",
            "Slower than base BLIP (but much more accurate)"
        ],
        "use_case": "Professional-grade technical image analysis"
    },

    "previous": {
        "name": "BLIP-image-captioning-base (Old)",
        "model_id": "Salesforce/blip-image-captioning-base",
        "strengths": [
            "Fast inference",
            "Good for general images"
        ],
        "limitations": [
            "Basic descriptions only",
            "Limited technical understanding"
        ],
        "use_case": "Upgraded to BLIP-2 - no longer current"
    },

    "recommended_upgrade": {
        "name": "BLIP-2 (Opt 2.7B)",
        "model_id": "Salesforce/blip2-opt-2.7b",
        "strengths": [
            "Much more detailed descriptions",
            "Better understanding of technical content",
            "Can answer questions about images (VQA)",
            "Handles complex scenes better"
        ],
        "limitations": [
            "Larger model (2.7B parameters)",
            "Slower inference",
            "Higher memory usage"
        ],
        "use_case": "Detailed technical plot analysis",
        "code_example": """
# To use BLIP-2 instead:
from transformers import Blip2Processor, Blip2ForConditionalGeneration

processor = Blip2Processor.from_pretrained("Salesforce/blip2-opt-2.7b")
model = Blip2ForConditionalGeneration.from_pretrained("Salesforce/blip2-opt-2.7b")

# For captioning:
inputs = processor(image, return_tensors="pt")
outputs = model.generate(**inputs, max_length=100)
caption = processor.decode(outputs[0], skip_special_tokens=True)

# For VQA (asking questions about images):
question = "What type of plot is this?"
inputs = processor(image, question, return_tensors="pt")
outputs = model.generate(**inputs)
answer = processor.decode(outputs[0], skip_special_tokens=True)
        """
    },

    "gpt4v_alternative": {
        "name": "GPT-4 Vision (API)",
        "model_id": "gpt-4-vision-preview",
        "strengths": [
            "Excellent understanding of technical diagrams",
            "Can explain complex plots in detail",
            "Understands MATLAB-specific visualizations",
            "Provides actionable insights"
        ],
        "limitations": [
            "Requires API key and internet",
            "Paid service",
            "Rate limits and costs"
        ],
        "use_case": "Professional-grade plot analysis",
        "code_example": """
# Using OpenAI GPT-4 Vision:
import openai
import base64

def analyze_plot_with_gpt4(image_path, question):
    with open(image_path, "rb") as image_file:
        base64_image = base64.b64encode(image_file.read()).decode('utf-8')

    response = openai.chat.completions.create(
        model="gpt-4-vision-preview",
        messages=[
            {
                "role": "user",
                "content": [
                    {"type": "text", "text": question},
                    {"type": "image_url", "image_url": {"url": f"data:image/png;base64,{base64_image}"}}
                ]
            }
        ],
        max_tokens=500
    )
    return response.choices[0].message.content
        """
    },

    "specialized_model": {
        "name": "Chart Understanding Models",
        "model_id": "Various (e.g., ChartOCR, ChartQA)",
        "strengths": [
            "Specifically trained on charts and plots",
            "Can extract data points and relationships",
            "Understands axis labels and legends"
        ],
        "limitations": [
            "May not handle non-chart images well",
            "Limited general image understanding"
        ],
        "use_case": "Data visualization analysis",
        "examples": [
            "microsoft/ChartQA",  # Chart question answering
            "google/deplot",      # Data extraction from plots
            "ChartOCR models"     # Chart-specific OCR
        ]
    }
}

def get_model_recommendation(current_performance):
    """
    Recommend next model based on current performance feedback.
    """
    if "basic descriptions" in current_performance or "not detailed enough" in current_performance:
        return VISION_MODEL_OPTIONS["recommended_upgrade"]

    if "needs better technical understanding" in current_performance:
        return VISION_MODEL_OPTIONS["gpt4v_alternative"]

    if "chart specific" in current_performance or "plot analysis" in current_performance:
        return VISION_MODEL_OPTIONS["specialized_model"]

    return VISION_MODEL_OPTIONS["current"]

def print_model_comparison():
    """Print a comparison of available vision models."""
    print("🔍 MATLAB RAG Vision Model Options")
    print("=" * 50)

    for key, model in VISION_MODEL_OPTIONS.items():
        print(f"\n🖼️  {model['name']}")
        print(f"   Model ID: {model['model_id']}")
        print(f"   Best for: {model.get('use_case', 'General use')}")

        print("   ✅ Strengths:")
        for strength in model['strengths']:
            print(f"      • {strength}")

        print("   ⚠️  Limitations:")
        for limitation in model['limitations']:
            print(f"      • {limitation}")

if __name__ == "__main__":
    print_model_comparison()

    print("\n" + "=" * 50)
    print("💡 Current BLIP Model Performance:")
    print("   • Good for basic plot detection")
    print("   • Recognizes 'plot with functions'")
    print("   • Needs more technical detail")
    print("\n🔄 Recommended Next Step: Upgrade to BLIP-2 for detailed descriptions")

    # Show upgrade path
    recommendation = get_model_recommendation("not detailed enough")
    print(f"\n🚀 Recommended Upgrade: {recommendation['name']}")
    print(f"   Why: {recommendation['use_case']}")
