requirements.txt
# Core RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1
langchain-openai>=0.0.5
langchain-ollama>=0.1.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
rank-bm25>=0.2.2

# Web server dependencies
flask>=2.3.0
flask-cors>=4.0.0
gradio>=4.0.0

# Vision and image processing dependencies
Pillow>=10.0.0
opencv-python-headless>=4.8.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
# BLIP-2 for advanced image understanding
# Note: BLIP-2 models are large (2.7B+ parameters)

# Additional dependencies
numpy>=1.24.0

